requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
mangum